Create README.md
This is for the Weather Research and Forecasting program (WRF) data.

This is a Java and Javascript program that runs as an applet within 
an HTML page. The applet is in Java so it can use existing Java 
NetCDF libraries. The Google Earth plugin can interact with 
Javascript, so that is used as the bridge.

The HTML page resides on the user's computer, as does any needed 
files, libraries, and programs.

The user selects a WRF related NetCDF file. The applet reads the 
variables within the file. The user selects one from a list and
chooses how to display it. The variable's data is converted to KML
and either saved to file or directly displayed onto the Google Earth
plugin within the HTML page. The user can also directly control a 
few of the Google Earth display options through the applet. The 
Google Earth plugin currently works only on Windows, but the applet
can write KML data to files without the plugin and the user can then
manually load the KML data into Google Earth program rather than the
plugin within the HTML page.

The user can select one or more grid cells within the data displayed
on the Google Earth plugin and change its values with those changes
written back to the NetCDF file. This is useful for changing input
data to WRF. If the plugin is not available or used, the Google 
Earth program can be used to accomplish similar functionality, but 
with more steps involved.